people_service/requirements.txt
kafka-python==2.0.2
fastapi[all]==0.100.1
python-dotenv==1.0.0
faker==19.2.0